V5 IndexSet: Tests and fixes.   - Changed to ulong IndexSet.   - Fixed lots of bugs. :/

+++ V5/V5.Native/ArraySearch.h
@@ -4,8 +4,8 @@ using namespace System;
 public ref class ArraySearch
 {
 public:	
-	static void AndWhereGreaterThan(array<Byte>^ set, Byte value, array<UInt32>^ matchVector);
+	static void AndWhereGreaterThan(array<Byte>^ set, Byte value, array<UInt64>^ matchVector);
 
-	static int Count(array<UInt32>^ matchVector);
+	static int Count(array<UInt64>^ matchVector);
 };
